pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "scCytoTrek"
version = "0.1.0"
description = "scRNA-seq analysis package with enhanced visualizations inspired by SeuratExtend."
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "scanpy~=1.9.0",
    "anndata>=0.8.0,<0.11.0",
    "numpy<2.0.0",
    "pandas",
    "scipy",
    "matplotlib",
    "seaborn",
    "harmonypy",
    "mudata",
    "scikit-learn",
    "networkx",
    "scvi-tools",
    "lifelines",
    "gseapy"
]
authors = [
    { name = "Yuefeng Wu@Westlake Univ, Kunliang Guan Lab" }
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-cov",
    "pre-commit",
    "black",
    "ruff"
]

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q"
testpaths = [
    "tests",
]
